third_party/WebKit/LayoutTests/svg/custom/use-nested-missing-target-removed-expected.txt

CONSOLE ERROR: line 7: Uncaught TypeError: Cannot set property 'id' of null
PASS
